code refactors to merge gremlin connector components into core/base
