+++ src/core/base/indicator.js
