+++ src/core/base/indicator.scss
