--- src/core/gremlin-connector/index.js
@@ -1,4 +1,0 @@
-import StatusMessageComponent from "./status";
-import ConnectionIndicatorComponent from "./indicator";
-
-export {StatusMessageComponent, ConnectionIndicatorComponent}
